add: `update` cli app for automatic package update
